requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0

# İsteğe bağlı — plaka tanıma yerel OCR (plaka_tanima.py):
# Pillow>=10.0.0
# pytesseract>=0.3.10
# (Sistemde tesseract-ocr da kurulu olmalı.)
# Harici API kullanılacaksa sadece .env içinde PLAKA_API_URL yeterli.
